2024-2025/PMP-42/bachelor-thesis/Andriy Shumskyy/requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
db==0.1.1
db-sqlite3==0.0.1
decorator==5.2.1
filelock==3.18.0
fire==0.7.0
fonttools==4.56.0
fsspec==2025.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
lmdb==1.6.2
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.3
sympy==1.13.1
termcolor==2.5.0
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.99
ultralytics-thop==2.0.14
urllib3==2.3.0
